use old version of stm32loder
